Created feed controller to download YouTube rss feed with video files as is commonly done for podcasts.
The code uses the existing YouTube rss feeds that can be accessed for channels and users and replaces the link to the YouTube video with a link to an enclosed video as specified by the format paramter. The replacement is done using the DOMDocument php class.

To generate such a feed for a channel, the channelid parameter must be passed.
To generate a feed for a user, the user parameter must be passed.
It is also necessary to pass the format parameter to determine the type of file that should be appended.